Do not write trailing NULL byte into user.shatag.sha256
Write 64 instead of 65 bytes.

The trailing NULL is useless bloat, Python shatag does not seem to write it,
and older cshatag versions should handle it just fine.

+++ tests/run_tests.sh
@@ -10,6 +10,12 @@ echo > foo.txt
 ../cshatag foo.txt
 ../cshatag foo.txt
 
+echo "*** Looking for NULL bytes (shouldn't find any)***"
+if getfattr -n user.shatag.sha256 foo.txt -e hex | grep 00 ; then
+	echo "error: NULL byte found"
+	exit 1
+fi
+
 echo "*** Testing modified empty file ***"
 echo > foo.txt
 ../cshatag foo.txt
